+++ app/app.go
@@ -780,6 +780,7 @@ func NewKYVEApp(
 			app.mm,
 			app.configurator,
 			app.AccountKeeper,
+			app.StakersKeeper,
 		),
 	)