+++ x/stakers/keeper/msg_server_create_staker_test.go
@@ -61,8 +61,10 @@ var _ = Describe("msg_server_create_staker.go", Ordered, func() {
 		Expect(staker.Commission).To(Equal(sdk.MustNewDecFromStr("0.2")))
 
 		Expect(staker.Moniker).To(BeEmpty())
-		Expect(staker.Logo).To(BeEmpty())
+		Expect(staker.Identity).To(BeEmpty())
 		Expect(staker.Website).To(BeEmpty())
+		Expect(staker.SecurityContact).To(BeEmpty())
+		Expect(staker.Details).To(BeEmpty())
 
 		Expect(valaccounts).To(BeEmpty())
 	})
@@ -99,8 +101,10 @@ var _ = Describe("msg_server_create_staker.go", Ordered, func() {
 		Expect(staker.Commission).To(Equal(types.DefaultCommission))
 
 		Expect(staker.Moniker).To(BeEmpty())
-		Expect(staker.Logo).To(BeEmpty())
+		Expect(staker.Identity).To(BeEmpty())
 		Expect(staker.Website).To(BeEmpty())
+		Expect(staker.SecurityContact).To(BeEmpty())
+		Expect(staker.Details).To(BeEmpty())
 
 		Expect(valaccounts).To(HaveLen(0))
 	})
@@ -153,8 +157,10 @@ var _ = Describe("msg_server_create_staker.go", Ordered, func() {
 		Expect(staker.Commission).To(Equal(types.DefaultCommission))
 
 		Expect(staker.Moniker).To(BeEmpty())
-		Expect(staker.Logo).To(BeEmpty())
+		Expect(staker.Identity).To(BeEmpty())
 		Expect(staker.Website).To(BeEmpty())
+		Expect(staker.SecurityContact).To(BeEmpty())
+		Expect(staker.Details).To(BeEmpty())
 
 		Expect(valaccounts).To(BeEmpty())
 	})
@@ -188,8 +194,10 @@ var _ = Describe("msg_server_create_staker.go", Ordered, func() {
 		Expect(staker.Commission).To(Equal(types.DefaultCommission))
 
 		Expect(staker.Moniker).To(BeEmpty())
-		Expect(staker.Logo).To(BeEmpty())
+		Expect(staker.Identity).To(BeEmpty())
 		Expect(staker.Website).To(BeEmpty())
+		Expect(staker.SecurityContact).To(BeEmpty())
+		Expect(staker.Details).To(BeEmpty())
 
 		Expect(valaccounts).To(BeEmpty())
 	})